Complete rewrite

+++ Assets/Scripts/Utils/ExtensionMethods.cs
@@ -2,7 +2,7 @@
 {
     public static class ExtensionMethods
     {
-        public static string ToHex(this short packetId)
+        public static string ToHex(this int packetId)
         {
             return "0x" + packetId.ToString("X");
         }
